+++ Core/BlazorMvvm.Core/Attributes/BlazorMvvmViewModelFactoryConstructorAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace BlazorMvvm
+{
+    [AttributeUsage(AttributeTargets.Constructor)]
+    public class BlazorMvvmViewModelFactoryConstructorAttribute : Attribute
+    {
+    }
+}