+++ Core/BlazorMvvm.Core/Mvvm/BlazorMvvmScopedCache.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace BlazorMvvm
+{
+    public class BlazorMvvmScopedCache
+    {
+        public Dictionary<Type, object> Cache { get; } = new();
+    }
+}