+++ Core/BlazorMvvm.Core/Mvvm/IBlazorMvvmViewModelFactory.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace BlazorMvvm
+{
+    public interface IBlazorMvvmViewModelFactory
+    {
+        IBlazorViewModel GetViewModel(Type viewModelType, IServiceProvider serviceProvider, BlazorMvvmScopedCache scopedCache);
+    }
+}