+++ Demo/BlazorMvvm.Demo/Program.cs
@@ -1,4 +1,4 @@
-using BlazorMvvm.Demo.JsHandlers;
+﻿using BlazorMvvm.Demo.JsHandlers;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using Microsoft.Extensions.DependencyInjection;
@@ -18,6 +18,8 @@ namespace BlazorMvvm.Demo
 
             builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
             builder.Services.AddSingleton<SvgHandler>();
+            
+            builder.Services.UseBlazorMvvmViewModelFactory();
 
             await builder.Build().RunAsync();
         }