+++ Tests/BlazorMvvm.Tests/Services/TestService.cs
@@ -1,0 +1,12 @@
+﻿namespace BlazorMvvm.Tests.Services
+{
+    public interface ITestService
+    {
+        string GetValue();
+    }
+
+    public class TestService : ITestService
+    {
+        public string GetValue() => "TestValue";
+    }
+}