+++ Tests/BlazorMvvm.Tests/ViewModel/LazyViewModel.cs
@@ -1,0 +1,13 @@
+﻿namespace BlazorMvvm.Tests.ViewModel
+{
+    [BlazorMvvmViewModel]
+    public class LazyViewModel : BlazorViewModel
+    {
+        public bool IsInitialized { get; private set; }
+
+        public LazyViewModel()
+        {
+            IsInitialized = true;
+        }
+    }
+}